Refactor python code into a single shared lambda package with multiple entrypoints.
