+++ .idea/vcs.xml
@@ -2,5 +2,6 @@
 <project version="4">
   <component name="VcsDirectoryMappings">
     <mapping directory="$PROJECT_DIR$" vcs="Git" />
+    <mapping directory="$PROJECT_DIR$/terraform/per_account/deployable/.terraform/modules/reference_accounts" vcs="Git" />
   </component>
 </project>
