--- lambda/cloudwatch_forwarder/Makefile
@@ -1,27 +1,0 @@
-.SHELL := /bin/sh
-.DEFAULT_GOAL := zip
-.PHONY = clean
-
-clean:
-	rm -rf __pycache__ .coverage *.zip *.egg-info .tox venv .pytest_cache htmlcov **/__pycache__ **/*.pyc .target setup.cfg
-
-test:
-	rm setup.cfg; tox
-
-target_dir:
-	rm -rf .target/
-	mkdir .target
-
-add_deps: target_dir
-	bash -c "echo -e '[install]\nprefix=\n' > setup.cfg"; python3 -m pip install -r requirements.txt -t .target
-
-copy_dir:
-	cp *.py .target/
-
-build: clean target_dir add_deps copy_dir
-
-run:
-	python generate_metric_alarms.py
-
-zip: build
-	cd .target; zip -X -9 ../cloudwatch_forwarder.zip -r .
