--- lambda/cloudwatch_forwarder/tox.ini
@@ -1,23 +1,0 @@
-[tox]
-envlist = 
-    py37
-skipsdist = true
-
-[testenv:py37]
-whitelist_externals =
-    python
-    pylint
-    flake8
-    pytest
-commands =
-    pylint tests/test_cloudwatch_forwarder.py
-    pylint tests/test_enrich.py
-    pylint cloudwatch_forwarder.py
-    pylint enrich.py
-    pylint -d W0613 lambda_handler.py
-    flake8 tests/test_cloudwatch_forwarder.py
-    flake8 tests/test_enrich.py
-    flake8 cloudwatch_forwarder.py
-    flake8 enrich.py
-    flake8 lambda_handler.py
-    pytest
