--- lambda/generate_metric_alarms/tox.ini
@@ -1,23 +1,0 @@
-[tox]
-envlist = 
-    py37
-skipsdist = true
-
-[testenv:py37]
-whitelist_externals =
-    python
-    pylint
-    flake8
-    pytest
-commands =
-    pylint tests/test_generate_metric_alarms.py
-    pylint generate_metric_alarms.py
-    pylint local_exceptions.py
-    pylint enrich.py
-    pylint -d W0613 lambda_handler.py
-    flake8 tests/test_generate_metric_alarms.py
-    flake8 generate_metric_alarms.py
-    flake8 local_exceptions.py
-    flake8 enrich.py
-    flake8 lambda_handler.py
-    pytest
