--- lambda/health_monitor/README.md
@@ -1,4 +1,0 @@
-# Cyber Security CloudWatch Config - Health Monitor Lambda
-
-Subscribes to an SNS notification and forwards to appropriate notification 
-channels. Slack / PD / Splunk / Some kind of dashboard app.
