--- lambda/health_monitor/requirements-dev.txt
@@ -1,5 +1,0 @@
--r requirements.txt
-pylint
-flake8
-pytest
-tox
