--- lambda/health_monitor/requirements.txt
@@ -1,2 +1,0 @@
-boto3==1.10.2
-addict==2.2.1
