--- lambda/health_monitor/tox.ini
@@ -1,17 +1,0 @@
-[tox]
-envlist = 
-    py37
-skipsdist = true
-
-[testenv:py37]
-whitelist_externals =
-    python
-    pylint
-    flake8
-    pytest
-commands =
-    pylint tests/test_health_monitor_lambda.py
-    pylint -d W0613 health_monitor_lambda.py
-    flake8 tests/test_health_monitor_lambda.py
-    flake8 health_monitor_lambda.py
-    pytest
