+++ lambda/health_package/Makefile
@@ -11,12 +11,14 @@ test:
 target_dir:
 	rm -rf .target/
 	mkdir .target
+	mkdir .target/components
 
 add_deps: target_dir
 	bash -c "echo -e '[install]\nprefix=\n' > setup.cfg"; python3 -m pip install -r requirements.txt -t .target
 
 copy_dir:
 	cp *.py .target/
+	cp components/*.py .target/components/
 
 build: clean target_dir add_deps copy_dir
 
@@ -24,4 +26,4 @@ run:
 	python generate_metric_alarms.py
 
 zip: build
-	cd .target; zip -X -9 ../cloudwatch_config.zip -r .
+	cd .target; zip -X -9 ../health_package.zip -r .
