+++ lambda/health_package/tox.ini
@@ -1,0 +1,47 @@
+[tox]
+envlist = 
+    py37
+skipsdist = true
+
+[testenv:py37]
+whitelist_externals =
+    python
+    pylint
+    flake8
+    pytest
+commands =
+    pylint tests/test_generate_metric_alarms.py
+    pylint tests/test_cloudwatch_forwarder.py
+    pylint tests/test_health_monitor_lambda.py
+    pylint tests/test_components_generic_helper.py
+    pylint tests/test_components_lambda_helper.py
+    pylint components/generic_helper.py
+    pylint components/firehose_helper.py
+    pylint components/kinesis_helper.py
+    pylint components/lambda_helper.py
+    pylint components/sqs_helper.py
+    pylint generate_metric_alarms.py
+    pylint cloudwatch_forwarder.py
+    pylint health_monitor_lambda.py
+    pylint local_exceptions.py
+    pylint -d C0103 enrich.py
+    pylint -d W0613 lambda_handler.py
+    pylint demo/set_alarm_state.py
+    flake8 tests/test_generate_metric_alarms.py
+    flake8 tests/test_cloudwatch_forwarder.py
+    flake8 tests/test_health_monitor_lambda.py
+    flake8 tests/test_components_generic_helper.py
+    flake8 tests/test_components_lambda_helper.py
+    flake8 components/generic_helper.py
+    flake8 components/firehose_helper.py
+    flake8 components/kinesis_helper.py
+    flake8 components/lambda_helper.py
+    flake8 components/sqs_helper.py
+    flake8 generate_metric_alarms.py
+    flake8 cloudwatch_forwarder.py
+    flake8 health_monitor_lambda.py
+    flake8 local_exceptions.py
+    flake8 enrich.py
+    flake8 lambda_handler.py
+    flake8 demo/set_alarm_state.py
+    pytest
